+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fixed every compat module to work for newer versions of the mods
+- Added mushroom variants